+++ src/stamps/appLog.js
@@ -1,6 +1,6 @@
-import stampit from 'stampit';
-import sessionStamp from './session';
-import { grab, post } from '../apiHelper';
+const stampit = require('stampit');
+const sessionStamp = require('./session');
+const { grab, post } = require('../apiHelper');
 
 const DEBUG = 'debug';
 const INFO = 'info';
@@ -103,4 +103,4 @@ const stamp = stampit()
 // Make sure we have the sessionStamp withSessionHandling method
 .compose(sessionStamp);
 
-export default stamp;
+module.exports = stamp;