+++ src/stamps/assets.js
@@ -1,6 +1,6 @@
-import stampit from 'stampit';
-import sessionStamp from './session';
-import { grab, grabRaw } from '../apiHelper';
+const stampit = require('stampit');
+const sessionStamp = require('./session');
+const { grab, grabRaw } = require('../apiHelper');
 
 const stamp = stampit()
 .methods({
@@ -21,4 +21,4 @@ const stamp = stampit()
 // Make sure we have the sessionStamp withSessionHandling method
 .compose(sessionStamp);
 
-export default stamp;
+module.exports = stamp;