+++ test/specs/client.spec.js
@@ -1,59 +1,65 @@
-import chai from 'chai';
-import factory from '../../src/index';
-
-const should = chai.should();
+const factory = require('../../src/index');
 
 describe('Appgrid Client creation', () => {
-  it('should throw when no param is passed', () => {
+  test('should throw when no param is passed', () => {
     const makeClient = () => factory();
 
-    makeClient.should.throw(Error);
+    expect(makeClient).toThrow(Error);
   });
 
-  it('should not throw when appKey and deviceId are passed', () => {
+  test('should not throw when appKey and deviceId are passed', () => {
     const makeClient = () => factory({
       appKey: '56ea6a370db1bf032c9df5cb',
       deviceId: 'gregTestingSDK'
     });
 
-    makeClient.should.not.throw(Error);
+    expect(makeClient).not.toThrow(Error);
   });
 
-  it('should throw when sessionKey is passed alone', () => {
+  test('should throw when sessionKey is passed alone', () => {
     const makeClient = () => factory({
       sessionKey: 'whatever'
     });
 
-    makeClient.should.throw(Error);
+    expect(makeClient).toThrow(Error);
   });
 
-  it('should not throw when sessionKey, appKey and deviceId are all passed', () => {
-    const makeClient = () => factory({
-      sessionKey: 'whatever',
-      appKey: '56ea6a370db1bf032c9df5cb',
-      deviceId: 'gregTestingSDK'
-    });
-
-    makeClient.should.not.throw(Error);
-  });
-
-  it('should throw when deviceId is passed without appKey and sessionKey', () => {
-    const makeClient = () => factory({
-      deviceId: 'stuff_here',
-    });
-
-    makeClient.should.throw(Error);
-  });
-
-  it('should not throw but generate a deviceId when appKey is passed without deviceId', () => {
-    let client;
-    const makeClient = () => {
-      client = factory({
+  test(
+    'should not throw when sessionKey, appKey and deviceId are all passed',
+    () => {
+      const makeClient = () => factory({
+        sessionKey: 'whatever',
         appKey: '56ea6a370db1bf032c9df5cb',
+        deviceId: 'gregTestingSDK'
       });
-    };
 
-    makeClient.should.not.throw(Error);
-    should.equal(true, client.props.config.deviceId && typeof client.props.config.deviceId === 'string');
-  });
+      expect(makeClient).not.toThrow(Error);
+    }
+  );
+
+  test(
+    'should throw when deviceId is passed without appKey and sessionKey',
+    () => {
+      const makeClient = () => factory({
+        deviceId: 'stuff_here',
+      });
+
+      expect(makeClient).toThrow(Error);
+    }
+  );
+
+  test(
+    'should not throw but generate a deviceId when appKey is passed without deviceId',
+    () => {
+      let client;
+      const makeClient = () => {
+        client = factory({
+          appKey: '56ea6a370db1bf032c9df5cb',
+        });
+      };
+
+      expect(makeClient).not.toThrow(Error);
+      expect(typeof client.props.config.deviceId === 'string').toBeTruthy();
+    }
+  );
 });